feat: add user management page and login navigation
